Move multiline import tests to importData and skip broken tests.
Multiline imports now use importData -> importFiles, so there is no point testing them in importText. Some test cases have broken, which was not noticed due to the lack of importData coverage.
